Addressing missing audit files
This additional code is for instances where a text audit file is missing, for example, when SurveyCTO Sync fails to download all of the text audit files. Whenever the code attempts to insheet or use a file that does not exist, the "cap" function is used, followed with code to continue on to the next file whenever it is not found (by using the _rc==601 if conditional statement).